feat: crud resposta

+++ src/main/java/br/com/forumhub/demo/service/TopicoService.java
@@ -1,21 +1,18 @@
 package br.com.forumhub.demo.service;
 
+import br.com.forumhub.demo.exceptions.UnauthorizedException;
 import br.com.forumhub.demo.dto.topico.TopicoRegisterDTO;
 import br.com.forumhub.demo.dto.topico.TopicoResponseDTO;
 import br.com.forumhub.demo.dto.topico.TopicoUpdateDTO;
 import br.com.forumhub.demo.model.entities.Topico;
 import br.com.forumhub.demo.repository.TopicoRepository;
-import br.com.forumhub.demo.repository.UsuarioRepository;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class TopicoService {
@@ -24,7 +21,7 @@ public class TopicoService {
     private TopicoRepository topicoRepository;
 
     @Autowired
-    private UsuarioRepository usuarioRepository;
+    private UsuarioService usuarioService;
 
     public TopicoResponseDTO criarTopico(TopicoRegisterDTO dto){
 
@@ -32,43 +29,46 @@ public class TopicoService {
             throw new IllegalArgumentException("Já existe um tópico com este título e mensagem.");
         }
 
-        var usuario = usuarioRepository.getReferenceById(dto.autorId());
+        var usuario = usuarioService.buscarUsuarioPorId(dto.autorId());
         var topico = new Topico(dto.titulo(), dto.mensagem(), usuario);
 
         topicoRepository.save(topico);
         return new TopicoResponseDTO(topico);
     }
 
-    public Page<TopicoResponseDTO> listarTopicos(Pageable pageable) {
-        return topicoRepository.findAll(pageable).map(TopicoResponseDTO::new);
+    public Page<Topico> listarTopicos(Pageable pageable) {
+        return topicoRepository.findAll(pageable);
     }
 
-    public Page<TopicoResponseDTO> buscarTopicosPorData(LocalDate data, Pageable pageable) {
-        return topicoRepository.findByDataCriacao(data, pageable).map(TopicoResponseDTO::new);
+    public Page<Topico> buscarTopicosPorData(LocalDate data, Pageable pageable) {
+        return topicoRepository.findByDataCriacao(data, pageable);
     }
 
-    public TopicoResponseDTO buscarTopicoPorId(Long id) {
-        return new TopicoResponseDTO(buscarPorId(id));
+    public Topico buscarTopicoPorId(Long id) {
+        return topicoRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Tópico com ID " + id + " não foi encontrado."));
     }
 
-    public TopicoResponseDTO atualizarTopico(Long id, @Valid TopicoUpdateDTO dto) {
-        var topico = buscarPorId(id);
+    public Topico atualizarTopico(Long id, @Valid TopicoUpdateDTO dto) {
+        var topico = buscarTopicoPorId(id);
 
         topico.setTitulo(dto.titulo());
         topico.setMensagem(dto.mensagem());
 
-        topicoRepository.save(topico);
-
-        return new TopicoResponseDTO(topico);
+        return topicoRepository.save(topico);
     }
 
-    public void deletarTopico(Long id) {
-        var topico = buscarPorId(id);
+    public void deletarTopico(Long usuarioId, Long topicoid) {
+        var usuario = usuarioService.buscarUsuarioPorId(usuarioId);
+        var topico = buscarTopicoPorId(topicoid);
+
+        if (!usuario.getId().equals(topico.getAutor().getId())) {
+            throw new UnauthorizedException("Usuário não autorizado a deletar este tópico.");
+        }
+
+
+
         topicoRepository.delete(topico);
     }
 
-    private Topico buscarPorId(Long id) {
-        return topicoRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Tópico com o id " + id + " não encontrado."));
-    }
 }
